Implement AttendeeActivity
