+++ app/src/androidTest/java/com/github/dedis/student20_pop/AttendeeActivityTest.java
@@ -1,14 +1,10 @@
-package com.github.dedis.student20_pop.ui;
+package com.github.dedis.student20_pop;
 
 import androidx.test.core.app.ActivityScenario;
-
-import com.github.dedis.student20_pop.R;
-import com.github.dedis.student20_pop.AttendeeActivity;
 
 import org.junit.Before;
 import org.junit.Test;
 
-import static androidx.test.core.app.ApplicationProvider.getApplicationContext;
 import static androidx.test.espresso.Espresso.onView;
 import static androidx.test.espresso.action.ViewActions.click;
 import static androidx.test.espresso.assertion.ViewAssertions.matches;
@@ -16,7 +12,7 @@ import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
 
-public class AttendeeFragmentTest {
+public class AttendeeActivityTest {
 
     @Before
     public void launchActivity() {
@@ -24,16 +20,22 @@ public class AttendeeFragmentTest {
     }
 
     @Test
-    public void onClickPropertiesTest() {
+    public void onClickHomeTest() {
+        onView(withId(R.id.tab_home)).perform(click());
+        onView(withId(R.id.fragment_home)).check(matches(isDisplayed()));
+    }
+
+    @Test
+    public void onClickPropertiesTest(){
         onView(withId(R.id.tab_properties)).perform(click());
         onView(withId(R.id.properties_view)).check(matches(isDisplayed()));
     }
 
     @Test
-    public void onClickListEventsTest() {
-        onView(withText(getApplicationContext().getString(R.string.past_events))).perform(click());
-        onView(withText(getApplicationContext().getString(R.string.present_events))).perform(click());
-        onView(withText(getApplicationContext().getString(R.string.future_events))).perform(click());
+    public void onClickListEventsTest(){
+        onView(withText("Past Events")).perform(click());
+        onView(withText("Present Events")).perform(click());
+        onView(withText("Future Events")).perform(click());
         onView(withId(R.id.event_layout)).check(matches(isDisplayed()));
     }
 }
