+++ app/src/androidTest/java/com/github/dedis/student20_pop/MainActivityTest.java
@@ -57,20 +57,4 @@ public class MainActivityTest {
         onView(withId(R.id.button_cancel_launch)).perform(click());
         onView(withId(R.id.fragment_home)).check(matches(isDisplayed()));
     }
-
-    @Ignore("Need to set testingAttendee to true in MainActivity")
-    @Test
-    public void onClickPropertiesTest(){
-        onView(withId(R.id.tab_properties)).perform(click());
-        onView(withId(R.id.properties_view)).check(matches(isDisplayed()));
-    }
-
-    @Ignore("Need to set testingAttendee to true in MainActivity")
-    @Test
-    public void onClickListEventsTest(){
-        onView(withText("Past Events")).perform(click());
-        onView(withText("Present Events")).perform(click());
-        onView(withText("Future Events")).perform(click());
-        onView(withId(R.id.event_layout)).check(matches(isDisplayed()));
-    }
 }
