+++ app/src/androidTest/java/com/github/dedis/student20_pop/OrganizerActivityTest.java
@@ -1,0 +1,26 @@
+package com.github.dedis.student20_pop;
+
+import androidx.test.core.app.ActivityScenario;
+
+import org.junit.Before;
+import org.junit.Test;
+
+import static androidx.test.espresso.Espresso.onView;
+import static androidx.test.espresso.action.ViewActions.click;
+import static androidx.test.espresso.assertion.ViewAssertions.matches;
+import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
+import static androidx.test.espresso.matcher.ViewMatchers.withId;
+
+public class OrganizerActivityTest {
+
+    @Before
+    public void launchActivity() {
+        ActivityScenario.launch(OrganizerActivity.class);
+    }
+
+    @Test
+    public void onClickHomeTest() {
+        onView(withId(R.id.tab_home)).perform(click());
+        onView(withId(R.id.fragment_home)).check(matches(isDisplayed()));
+    }
+}
